Refactor `optimize.rs` into an `ir_generation` module.
