+++ sway-core/src/semantic_analysis.rs
@@ -1,6 +1,5 @@
 //! Type checking for Sway.
 pub mod ast_node;
-mod const_eval;
 mod module;
 pub mod namespace;
 mod node_dependencies;
